!refactor: Move get-capabilities files
