+++ projects/hslayers/src/common/get-capabilities/arcgis-get-capabilities.service.ts
