+++ projects/hslayers/src/common/get-capabilities/get-capabilities.module.ts
@@ -1,8 +1,8 @@
-import {HsArcgisGetCapabilitiesService} from '../arcgis/get-capabilities.service';
+import {HsArcgisGetCapabilitiesService} from './arcgis-get-capabilities.service';
 import {HsDimensionService} from '../dimension.service';
-import {HsWfsGetCapabilitiesService} from '../wfs/get-capabilities.service';
-import {HsWmsGetCapabilitiesService} from '../wms/get-capabilities.service';
-import {HsWmtsGetCapabilitiesService} from '../wmts/get-capabilities.service';
+import {HsWfsGetCapabilitiesService} from './wfs-get-capabilities.service';
+import {HsWmsGetCapabilitiesService} from './wms-get-capabilities.service';
+import {HsWmtsGetCapabilitiesService} from './wmts-get-capabilities.service';
 import {NgModule} from '@angular/core';
 
 @NgModule({
