+++ projects/hslayers/src/common/get-capabilities/wms-get-capabilities-response.interface.ts
