+++ projects/hslayers/src/common/get-capabilities/wms-get-capabilities.service.ts
